requirements.txt
streamlit==1.25.0
pandas==2.2.2
plotly==5.15.0
facebook-business==17.0.2
python-dotenv==1.0.0
openai==0.28.0
xlsxwriter==3.1.2
numpy==1.26.4 